part4 - pieniöä korjauksia kaipaa

+++ part4/controllers/blogs.js
@@ -24,15 +24,13 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const body = request.body 
-
+  const body = request.body
+  
   const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })
   }
   const user = await User.findById(decodedToken.id)
-
-  console.log('user-body', user)
   
   const blog = new Blog({
     title: body.title,
@@ -50,7 +48,17 @@ blogsRouter.post('/', async (request, response) => {
 )
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
+  
+  const blog = await Blog.findById(request.params.id)
+  const user = request.body.user
+  console.log(user)
+  if ( blog.user.toString() === user.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'invalid signature for the delete' })
+  }
+
   response.status(204).end()
 })
